Separate class for OSL controller
