+++ .gitmodules
@@ -1,6 +1,7 @@
 [submodule "myosuite/simhive/myo_sim"]
 	path = myosuite/simhive/myo_sim
 	url = https://github.com/MyoHub/myo_sim.git
+	branch = dev
 [submodule "myosuite/simhive/furniture_sim"]
 	path = myosuite/simhive/furniture_sim
 	url = https://github.com/vikashplus/furniture_sim.git
